Moving Result to inside of ImplementationLoader
